# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:9000
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
    Bearer:
        type: apiKey
        name: Authorization
        in: header
paths:
...
  /savetimesheet:
    # binds a127 app logic to a route
    x-swagger-router-controller: save_timesheet
    post:
      description: Returns 'Timesheet' to the caller
      # used as the method name of the controller
      operationId: saveTimesheet
      security:
        - Bearer: []
      parameters:
        - name: "body"
          in: "body"
          description: The email of the person to whom to say hello
          required: false
          schema:
            $ref: "#/definitions/savetimesheet" 
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"            
...
  /saveTasks:
    # binds a127 app logic to a route
    x-swagger-router-controller: task_add
    post:
      description: Returns 'task_add' to the caller
      # used as the method name of the controller
      operationId: saveTasks
      security:
        - Bearer: []
      parameters:
        - name: "body"
          in: "body"
          description: The email of the person to whom to say hello
          required: false
          schema:
            $ref: "#/definitions/taskadds" 
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"            
  /FetchAllTasks:
    # binds a127 app logic to a route
    x-swagger-router-controller: task_add
    get:
      description: Returns 'FetchAllTasks' to the caller
      # used as the method name of the controller
      operationId: FetchAllTasks
      security:
        - Bearer: []
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /getTimecardData:
    # binds a127 app logic to a route
    x-swagger-router-controller: save_timesheet
    get:
      description: Returns 'rpttimecard' to the caller
      # used as the method name of the controller
      operationId: fetchTimeCardData
      security:
        - Bearer: []
      parameters:
        - name: userID
          in: query
          description: The email of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /fetchFullTaskDatas:
    # binds a127 app logic to a route
    x-swagger-router-controller: task_add
    get:
      description: Returns 'getFullTaskDatas' to the caller
      # used as the method name of the controller
      operationId: getFullTaskDatas
      security:
        - Bearer: []
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /getTimeTrackerDetails:
    # binds a127 app logic to a route
    x-swagger-router-controller: save_timesheet
    get:
      description: Returns 'rpttimecard' to the caller
      # used as the method name of the controller
      operationId: getTimeTrackerDetails
      security:
        - Bearer: []
      parameters:
        - name: timetrackerid
          in: query
          description: The email of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /deleteTimetracker:
    # binds a127 app logic to a route
    x-swagger-router-controller: save_timesheet
    delete:
      description: Returns 'rpttimecard' to the caller
      # used as the method name of the controller
      operationId: deleteTimetracker
      security:
        - Bearer: []
      parameters:
        - name: timetrackerid
          in: query
          description: The email of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /deleteTasklist:
    # binds a127 app logic to a route
    x-swagger-router-controller: task_add
    delete:
      description: Returns 'tasks' to the caller
      # used as the method name of the controller
      operationId: deleteTasklist
      security:
        - Bearer: []
      parameters:
        - name: taskid
          in: query
          description: The email of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /fetchTaskByID:
    # binds a127 app logic to a route
    x-swagger-router-controller: task_add
    get:
      description: Returns 'tasks' to the caller
      # used as the method name of the controller
      operationId: fetchTaskByID
      security:
        - Bearer: []
      parameters:
        - name: taskid
          in: query
          description: The email of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /fetchAllTaskbyname:
    # binds a127 app logic to a route
    x-swagger-router-controller: task_add
    get:
      description: Returns 'fetchAllTaskbyname' to the caller
      # used as the method name of the controller
      operationId: fetchAllTaskbyname
      security:
        - Bearer: []
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /updateTimesheet:
    # binds a127 app logic to a route
    x-swagger-router-controller: save_timesheet
    put:
      description: Returns 'Timesheet' to the caller
      # used as the method name of the controller
      operationId: updateTimecard
      security:
        - Bearer: []
      parameters:
        - name: "timesheetID"
          in: "body"
          description: The email of the person to whom to say hello
          required: false
          schema:
            $ref: "#/definitions/updateTimecard"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"            
...
  /getTimeTrackFull:
    # binds a127 app logic to a route
    x-swagger-router-controller: save_timesheet
    get:
      description: Returns 'rpttimecard' to the caller
      # used as the method name of the controller
      operationId: getTimeTrackFull
      security:
        - Bearer: []
      parameters:
        - name: userID
          in: query
          description: The email of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /fullTimesheet:
    # binds a127 app logic to a route
    x-swagger-router-controller: save_timesheet
    get:
      description: Returns 'rpttimecard' to the caller
      # used as the method name of the controller
      operationId: fullTimesheet
      security:
        - Bearer: []
      parameters:
        - name: userID
          in: query
          description: The email of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
...
  /fullTimesheetAdmin:
    # binds a127 app logic to a route
    x-swagger-router-controller: save_timesheet
    get:
      description: Returns 'rpttimecard' to the caller
      # used as the method name of the controller
      operationId: fullTimesheetAdmin
      security:
        - Bearer: []
      parameters:
        - name: userID
          in: query
          description: The email of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"   
  /fetchAdminUsersById:
    # binds a127 app logic to a route
    x-swagger-router-controller: save_timesheet
    get:
      description: Returns 'users' to the caller
      # used as the method name of the controller
      operationId: fullTimesheet
      security:
        - Bearer: []
      parameters:
        - name: userID
          in: query
          description: The email of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"         
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  savetimesheet:
    type: "object"
    required:
      - "project_id"
      - "task_id"
      - "notes"
      - "entertime"
    properties:
      project_id:
        type: "string"
        example: "project_name"
      task_id:
        type: "string"
        example: "task_name"
      notes:
        type: "string"
        example: "notes"
      entertime:
        type: "object"
        example: "entertime"
...
  taskadds:
    type: "object"
    required:
      - "project_id"
      - "task_name"
    properties:
      project_id:
        type: "string"
        example: "project_name"
      task_name:
        type: "string"
        example: "task_name"
...
  updateTimecard:
    type: "object"
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string